+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Baseline Builder
 
-## 1.0.8
+## 1.1.0
 - CI: Switch to `svenstaro/upload-release-action@v2`
 - Validation: Ensure Installomator labels are valid
 - Build: Add support for mobileconfig-based inputs